refactor to multiple files for each view
